requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
aiohappyeyeballs @ file:///C:/b/abs_e9fm5mk6ak/croot/aiohappyeyeballs_1734469429350/work
aiohttp @ file:///C:/b/abs_d4_gdjffav/croot/aiohttp_1734687146788/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
alembic==1.15.2
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_d8yc2t_bxy/croot/async-timeout_1732663677668/work
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
casadi==3.7.0
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_f9ihnt72pu/croot/click_1698129847492/work
cloudpickle==3.1.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
colorlog==6.9.0
contourpy==1.3.0
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cycler==0.12.1
dill==0.3.9
do_mpc==4.6.5
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
fonttools==4.57.0
frozenlist @ file:///C:/b/abs_06ctmb1zeo/croot/frozenlist_1730903113463/work
fsspec==2025.3.2
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-auth @ file:///C:/b/abs_86jbgkfwa2/croot/google-auth_1741277194305/work
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
greenlet==3.1.1
grpcio @ file:///C:/ci/grpcio_1637590978642/work
gym==0.26.2
gym-notices==0.0.8
h5py @ file:///C:/b/abs_9cz_vyf1xj/croot/h5py_1737051135810/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
importlib_resources==6.5.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1637159014053/work/keras-2.6.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keras-rl==0.4.2
kiwisolver==1.4.7
Mako==1.3.9
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.4
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
msgpack==1.1.0
multidict @ file:///C:/b/abs_19e3ubo2ew/croot/multidict_1730905504444/work
namex==0.0.8
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.26.4
oauthlib @ file:///C:/b/abs_44riqpe5gg/croot/oauthlib_1679489644705/work
opensim==4.5
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
optree==0.15.0
optuna==4.2.1
osim-rl @ git+https://github.com/stanfordnmbl/osim-rl.git@610b95cf0c4484f1acecd31187736b0113dcfb73
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pooch @ file:///C:/b/abs_7acoiuwpey/croot/pooch_1731007069545/work
propcache @ file:///C:/b/abs_d6o8xbonwb/croot/propcache_1732304003668/work
protobuf==6.30.2
pyarrow==19.0.1
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydy==0.7.1
Pygments==2.19.1
PyJWT @ file:///C:/b/abs_02n5wes3j9/croot/pyjwt_1736178904450/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.2.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
ray==2.44.1
redis==5.2.1
referencing==0.36.2
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-oauthlib @ file:///C:/b/abs_faw0a7_wag/croot/requests-oauthlib_1720615036938/work
rich==14.0.0
rpds-py==0.24.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.40
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.6.2.2
tensorflow==2.6.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
timeout-decorator==0.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt @ file:///C:/b/abs_68gzx33lzh/croot/wrapt_1736543196433/work
yarl @ file:///C:/b/abs_281qby2vim/croot/yarl_1732546854547/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work

